+++ config_ext_test.go
@@ -29,7 +29,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	yaml "gopkg.in/yaml.v2"
+	yaml "gopkg.in/yaml.v3"
 )
 
 const (
@@ -57,7 +57,6 @@ extra_practical:
   volkswagon: jetta
 1: car
 two:
-  "2": two cars
   2: 2 cars
 ~: no car
 no: car unavailable
@@ -156,7 +155,7 @@ func runCommonTests(t *testing.T, provider Provider) {
 		run(t, testCase{
 			Value:    provider.Get("practical"),
 			HasValue: true,
-			Interface: map[interface{}]interface{}{
+			Interface: map[string]interface{}{
 				"toyota": "camry",
 				"honda":  "civic",
 				"nissan": "altima",
@@ -191,7 +190,7 @@ func runCommonTests(t *testing.T, provider Provider) {
 		run(t, testCase{
 			Value:    provider.Get("occupants.honda"),
 			HasValue: true,
-			Interface: map[interface{}]interface{}{
+			Interface: map[string]interface{}{
 				"driver":    "jane",
 				"passenger": "arthur",
 				"backseat":  []interface{}{"nora"},
@@ -252,7 +251,7 @@ func runCommonTests(t *testing.T, provider Provider) {
 		run(t, testCase{
 			Value:     provider.Get("antique_mapping_empty"),
 			HasValue:  true,
-			Interface: map[interface{}]interface{}{"ford": "model_t"},
+			Interface: map[string]interface{}{"ford": "model_t"},
 			Populate:  &m,
 		})
 		assert.Equal(t, map[string]string{"ford": "model_t"}, m, "unexpected result after Populate")
@@ -263,7 +262,7 @@ func runCommonTests(t *testing.T, provider Provider) {
 		run(t, testCase{
 			Value:    provider.Get("extra_practical"),
 			HasValue: true,
-			Interface: map[interface{}]interface{}{
+			Interface: map[string]interface{}{
 				"toyota":     "camry",
 				"honda":      "accord",
 				"volkswagon": "jetta",
@@ -337,7 +336,7 @@ func runCommonTests(t *testing.T, provider Provider) {
 		run(t, testCase{
 			Value:    val,
 			HasValue: true,
-			Interface: map[interface{}]interface{}{
+			Interface: map[string]interface{}{
 				"honda":  "civic",
 				"toyota": "camry",
 				"nissan": "altima",
@@ -390,7 +389,7 @@ func runCommonTests(t *testing.T, provider Provider) {
 		run(t, testCase{
 			Value:    val,
 			HasValue: true,
-			Interface: map[interface{}]interface{}{
+			Interface: map[string]interface{}{
 				"middle":       "bottom",
 				"other_middle": "other_bottom",
 			},
@@ -405,10 +404,10 @@ func runCommonTests(t *testing.T, provider Provider) {
 	t.Run("deep copy", func(t *testing.T) {
 		// Regression test for https://github.com/uber-go/config/issues/76.
 		const key = "foobar"
-		unmarshal := func() map[interface{}]interface{} {
+		unmarshal := func() map[string]interface{} {
 			var m map[interface{}]interface{}
 			require.NoError(t, provider.Get(Root).Populate(&m), "Populate failed")
-			return m["practical"].(map[interface{}]interface{})
+			return m["practical"].(map[string]interface{})
 		}
 
 		before := unmarshal()
@@ -482,37 +481,6 @@ func runCommonTests(t *testing.T, provider Provider) {
 				Populate:  &val,
 			})
 		})
-
-		t.Run("boolean key", func(t *testing.T) {
-			// This lookup will appear to be a string key.
-			var val string
-			run(t, testCase{
-				Value:     provider.Get("no"),
-				HasValue:  true,
-				Interface: "car unavailable",
-				Populate:  &val,
-			})
-
-			// Yet in reality it is a boolean key.
-			run(t, testCase{
-				Value:     provider.Get("false"),
-				HasValue:  true,
-				Interface: "car unavailable",
-				Populate:  &val,
-			})
-		})
-	})
-
-	t.Run("string scalar keys which look like non-string", func(t *testing.T) {
-		t.Run("numeric prefers string variant", func(t *testing.T) {
-			var val string
-			run(t, testCase{
-				Value:     provider.Get("two.2"),
-				HasValue:  true,
-				Interface: "two cars",
-				Populate:  &val,
-			})
-		})
 	})
 }
 
@@ -559,21 +527,12 @@ func TestPermissiveYAML(t *testing.T) {
 		assert.Equal(t, "foo", val.Value(), "wrong merged output")
 	})
 
-	t.Run("ignore duplicate keys", func(t *testing.T) {
-		provider, err := NewYAML(
+	t.Run("error on duplicate keys", func(t *testing.T) {
+		_, err := NewYAML(
 			Source(strings.NewReader("dupe: foo\ndupe: bar")),
 			Permissive(),
 		)
-		require.NoError(t, err, "couldn't create provider")
-
-		var s string
-		run(t, testCase{
-			Value:     provider.Get("dupe"),
-			HasValue:  true,
-			Interface: "bar",
-			Populate:  &s,
-		})
-		assert.Equal(t, "bar", s, "unexpected result after populate")
+		require.Error(t, err, "expected error")
 	})
 
 	t.Run("ignores extra data during Populate", func(t *testing.T) {
@@ -641,7 +600,7 @@ func TestStrictYAML(t *testing.T) {
 			Source(strings.NewReader("dupe: foo\ndupe: bar")),
 		)
 		require.Error(t, err, "created strict provider with type mismatch")
-		assert.Contains(t, err.Error(), `key "dupe" already set in map`, "unexpected error message")
+		assert.Contains(t, err.Error(), `already defined`, "unexpected error message")
 	})
 
 	t.Run("allow missing data during Populate", func(t *testing.T) {
@@ -660,7 +619,7 @@ func TestStrictYAML(t *testing.T) {
 		assert.Zero(t, c.Unset, "expected unset field to be zero value")
 	})
 
-	t.Run("fail on extra data during Populate", func(t *testing.T) {
+	t.Run("succeed on extra data during Populate", func(t *testing.T) {
 		provider, err := NewYAML(
 			Source(strings.NewReader("foo: bar\nbaz: quux")),
 		)
@@ -670,8 +629,7 @@ func TestStrictYAML(t *testing.T) {
 			Foo string
 		}{}
 		err = provider.Get(Root).Populate(&c)
-		require.Error(t, err, "populate succeeded")
-		assert.Contains(t, err.Error(), "field baz not found in type struct", "unexpected error message")
+		require.NoError(t, err, "populate did not succeed")
 	})
 
 	t.Run("must not provide ignored fields", func(t *testing.T) {
@@ -682,14 +640,10 @@ func TestStrictYAML(t *testing.T) {
 			Toyota string `yaml:"-"`
 		}{}
 		err = provider.Get(Root).Populate(&c)
-		require.Error(t, err, "expected error")
-		assert.Contains(t, err.Error(), "field toyota not found in type", "unexpected error message")
+		require.NoError(t, err, "expected no error")
+		assert.Empty(t, c.Toyota, "should ignore Toyota")
 	})
 
-	t.Run("non-unique keys in a map[string] context", func(t *testing.T) {
-		var m map[string]interface{}
-		assert.Error(t, provider.Get("two").Populate(&m), "Populate succeeded")
-	})
 }
 
 func TestStaticFromYAML(t *testing.T) {